Move reused styles to HeaderStyles.tsx and move NavigationLink to own component
